modified file structure to match jupyter repo
